// src/pages/table-test/types.ts
export type User = {
    name: string
    age: number
    email: string
    gender: "男" | "女" | "其他"
    registrationDate: string
    status: "活跃" | "非活跃"
    profilePicture: string
    phone: string
    address: string
    company: string
    department: string
    position: string
    salary: number
    level: string
    skills: string[]
    education: string
    graduateSchool: string
    major: string
    entryDate: string
    birthDate: string
}